requirements.txt
accelerate
anthropic
hydra-core
hydra-submitit-launcher
jinja2
openai
outlines<0.0.39
pandas
pillow
pydantic
randomname
sentencepiece
timm
torch
torchvision
transformers
vertexai
wandb
wikipedia-api
pybraille
qwen-vl-utils